Fix backport and demos to be compilable with Xcode 12.5 / Swift 5.4

+++ Backports/ProgressView.swift
@@ -136,7 +136,7 @@ private struct LinearProgressView: View {
                 .overlay(
                     RoundedRectangle(cornerRadius: 2)
                         .foregroundColor(.accentColor)
-                        .frame(width: proxy.size.width * fractionCompleted),
+                        .frame(width: proxy.size.width * CGFloat(fractionCompleted)),
                     alignment: .leading
                 )
         }
